moves schema stuff to schema files
