+++ storage_bucket/aws/create.py
@@ -1,16 +1,7 @@
-from typing import Literal, Union
-
-from storage_bucket.create_bucket import BaseCreateBucket
+from storage_bucket.aws.schema import AWSCreateBucket
 from storage_bucket.enums import Platform
 from storage_bucket.registries import register_create_bucket
 from storage_bucket.aws.client import aws_get_client
-
-
-class AWSCreateBucket(BaseCreateBucket):
-    """Google Cloud Platform create bucket parameters."""
-
-    storage_class: Union[Literal['STANDARD'], Literal['TEST']]
-    location: str
 
 
 @register_create_bucket(platform=Platform.aws)
@@ -18,7 +9,7 @@ def aws_create_bucket(
     create_params: AWSCreateBucket,
 ) -> int:
     """Create a gcp storage bucket."""
-    client = aws_get_client()
     print("in aws create bucket")
+    client = aws_get_client()
 
     return 1
