+++ storage_bucket/create_bucket.py
@@ -1,26 +1,20 @@
-from typing import TYPE_CHECKING, Any, Literal, Union, overload
+from typing import TYPE_CHECKING, Literal, NoReturn, Union, overload
 
-from attr import dataclass
-
+from storage_bucket.aws.schema import AWSCreateBucket
 from storage_bucket.enums import Platform
+from storage_bucket.gcp.schema import GCPCreateBucket
 from storage_bucket.registries import get_create_bucket
+from storage_bucket.schema import BaseCreateBucket
 
 if TYPE_CHECKING:
     from google.cloud.storage.bucket import Bucket
-
-
-@dataclass
-class BaseCreateBucket(object):
-    """Base shared create bucket parameters."""
-
-    name: str
 
 
 @overload
 def create_bucket(
     *,
     platform: Literal[Platform.gcp],
-    create_params: BaseCreateBucket,
+    create_params: GCPCreateBucket,
 ) -> 'Bucket':
     """Return strings when platform is gcp."""
 
@@ -29,9 +23,9 @@ def create_bucket(
 def create_bucket(
     *,
     platform: Literal[Platform.aws],
-    create_params: BaseCreateBucket,
-) -> int:
-    """Return int when platform is aws."""
+    create_params: AWSCreateBucket,
+) -> NoReturn:
+    """No good return if create_params is not GCPCreateBucket."""
 
 
 @overload
@@ -39,14 +33,14 @@ def create_bucket(
     *,
     platform: Platform,
     create_params: BaseCreateBucket,
-) -> Any:
-    """Not sure what we return if platform is not provided."""
+) -> NoReturn:
+    """No good return if create_params is not GCPCreateBucket."""
 
 
 def create_bucket(
     *,
     platform: Platform,
     create_params: BaseCreateBucket,
-) -> Union[int, str, Any]:
+) -> Union['Bucket', int]:
     """Get correct create bucket function and run it."""
     return get_create_bucket(platform)(create_params=create_params)
