+++ storage_bucket/schema.py
@@ -1,0 +1,9 @@
+from attr import dataclass
+
+
+
+@dataclass
+class BaseCreateBucket(object):
+    """Base shared create bucket parameters."""
+
+    name: str
